pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["rattus-aristarchus <lankin.m.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.4.3"
selenium = "4.11.2"
selene = "2.0.0rc4"
python-dotenv = "0.20.0"
allure-pytest = "2.13.2"
requests = "2.31.0"
lxml = "4.9.3"
feedparser = "6.0.10"
Appium-Python-Client = "2.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--clean-alluredir",
    "--alluredir=tests/allure-results"
]
